requirements.txt
Flask
requests
PyJWT
tempmail-lol
fastapi
uvicorn